tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/101.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: defer script after initial onload event</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<body>

	<div id="log">FAILED (This TC requires JavaScript enabled)</div>

	<script>
          var t = async_test();
          onload = t.step_func(
            function() {
              document.open();
              document.write("<title> scheduler: defer script after initial onload event</title><script src='/resources/testharness.js'><\/script><script src='/resources/testharnessreport.js'><\/script><script src='testlib/testlib.js'><\/script><div id='log'>document.written content</div><script>log('inline script #1'); t = async_test();<\/script><script src='scripts/include-1.js'><\/script><script defer src='scripts/include-2.js'><\/script>");
              document.close();
              //Note that the *window* object has changed but the *global scope* of the script has not.
              window.setTimeout(
                function() {
                  window.t.step(
                    function() {
                      window.assert_array_equals(window.eventOrder, ['inline script #1', 'external script #1', 'external script #2']);
                      window.t.done();
                })}, 1000);
            });
        </script>
</body>
</html>
